:recycle: split event header
